Cleaning up the code, and extracting out functionality.
