+++ go.mod
@@ -2,7 +2,6 @@ module github.com/Natasha-H-S/Go-Book-API
 
 go 1.18
 
-require (
-	github.com/go-sql-driver/mysql v1.6.0 // indirect
-	upper.io/db.v3 v3.8.0+incompatible // indirect
-)
+require upper.io/db.v3 v3.8.0+incompatible
+
+require github.com/go-sql-driver/mysql v1.6.0 // indirect
